Add the first case
The first case consist in two friends a number of steps apart from each other.
Everytime one takes a step closer to the other one, his tiredness increases by
a coeficient. The coeficient increases by one at every step. The objective is
to make they reach each other, with the smallest tiredness possible. To do so,
I opted to make each friend take a step alternated, making their coeficient
increase the least possible.
I chose this approach because I thought it was easier to read, since the case
has a story, and tried to make it the closest posible to the real world logic.